+++ apps/weather-forecast/src/app/app.component.scss
@@ -9,17 +9,6 @@
 	margin: 50px auto;
 }
 
-.search {
-}
-
-.mode {
-	display: block;
-	margin-bottom: 16px;
-}
-
-.table {
-}
-
 .gutter-left {
 	margin-left: 9px;
 }
