+++ apps/weather-forecast/src/app/components/search/search.component.html
@@ -1,5 +1,5 @@
 <form [formGroup]="form" (submit)="onSearch()" class="search">
-	<input type="text" formControlName="search" class="search__input" />
+	<input type="text" placeholder="Search city" formControlName="search" class="search__input" />
 	<button type="button" (click)="onSearch()" class="search__button" [disabled]="form.invalid || isLoading">
 		Search
 	</button>
