+++ apps/weather-forecast/src/app/pages/main/main.component.scss
@@ -1,0 +1,12 @@
+:host {
+	display: block;
+	font-family: sans-serif;
+	min-width: 300px;
+	max-width: 600px;
+	margin: 50px auto;
+}
+
+.mode {
+	display: block;
+	margin-bottom: 16px;
+}
